+++ doc-app/doc-ui-common/client/markdown/toc/TableOfContents.control.ts
@@ -1,6 +1,6 @@
 import * as rx from 'rxjs';
 import * as op from 'rxjs/operators';
-import {ReactorComposite, ActionTableDataType} from '@wfh/reactivizer';
+import {ReactorComposite, ActionTableDataType, actionRelatedToPayload} from '@wfh/reactivizer';
 import {useAppLayout} from '../../components/appLayout.control';
 import {markdownsControl} from '../markdownSlice';
 import {TOC} from '../../../isom/md-types';
@@ -10,24 +10,19 @@ export type ItemState = {
   expanded?: boolean;
   highlighted?: boolean;
   children?: string[]; // hashes
+  textDom?: HTMLDivElement;
 } & Omit<TOC, 'children'>;
-
-export type TocUIState = {
-  topLevelItems: string[];
-  positionFixed: boolean;
-  itemByHash?: Map<string, ItemState>;
-};
 
 export type TocUIActions = {
   setLayoutControl(layout: NonNullable<ReturnType<typeof useAppLayout>>): void;
   setDataKey(key: string): void;
+  setMarkdownBodyRef(key: string, dom: HTMLDivElement): void;
   expand(id: string, isExpand: boolean): void;
   clicked(id: string): void;
   onPlaceHolderRef(ref: HTMLDivElement | null): void;
   onContentDomRef(ref: HTMLDivElement | null): void;
   onContentScroll(): void;
-  showPopup(): void;
-  hidePopup(): void;
+  togglePopup(isOn: boolean): void;
 };
 
 type TocUIEvents = {
@@ -38,10 +33,15 @@ type TocUIEvents = {
 
   loadRawItem(toc: TOC): void;
   itemById(map: Map<string, ItemState>): void;
+  togglePopup: TocUIActions['togglePopup'];
+  togglePopupClassName(cln: string): void;
 };
 
-const tocInputTableFor = ['expand', 'setDataKey', 'onPlaceHolderRef', 'onContentDomRef', 'showPopup', 'hidePopup'] as const;
-const tocOutputTableFor = ['changeFixPosition', 'topLevelItemIdsUpdated', 'itemById'] as const;
+const tocInputTableFor = ['expand', 'setDataKey', 'onPlaceHolderRef', 'onContentDomRef', 'togglePopup'] as const;
+const tocOutputTableFor = ['changeFixPosition', 'topLevelItemIdsUpdated', 'itemById', 'togglePopup',
+  'togglePopupClassName'] as const;
+
+export type TocUIEventTable = ActionTableDataType<TocUIEvents, typeof tocOutputTableFor>;
 
 export function createControl(uiDirtyCheck: (immutableObj: any) => any) {
   const composite = new ReactorComposite<TocUIActions, TocUIEvents, typeof tocInputTableFor, typeof tocOutputTableFor>({
@@ -65,8 +65,9 @@ export function createControl(uiDirtyCheck: (immutableObj: any) => any) {
   ));
 
   r('setDataKey -> loadRowItem', i.pt.setDataKey.pipe(
-    rx.switchMap(([m, key]) => markdownsControl.outputTable.l.contents.pipe(
-      op.map(([, contents]) => contents[key]),
+    rx.switchMap(([m, key]) => markdownsControl.outputTable.l.htmlDone.pipe(
+      rx.filter(([, key0]) => key0 === key),
+      op.map(([, , contents]) => contents),
       op.distinctUntilChanged(),
       op.filter(data => data != null),
       op.take(1),
@@ -85,27 +86,110 @@ export function createControl(uiDirtyCheck: (immutableObj: any) => any) {
     ))
   ));
 
-  r('set map for itemState', o.pt.itemUpdated.pipe(
+  r('itemUpdated -> itemById, itemDomRefHandlers', o.pt.itemUpdated.pipe(
     rx.withLatestFrom(outputTable.l.itemById),
     rx.tap(([[, it], [, map]]) => {
       map.set(it.id, it);
     })
   ));
 
-  r('layout onTopAppBarScrollChange -> changeFixPosition', i.pt.setLayoutControl.pipe(
-    rx.switchMap(([, layout]) => layout.outputTable.l.onTopAppBarScrollChange),
-    rx.map(([, outOfViewPort]) => {
-      o.dp.changeFixPosition(outOfViewPort);
-    })
+  r('setDataKey, onContentDomRef, topLevelItemIdsUpdated -> scan HTML for heads (itemById.textDom)',
+    rx.combineLatest([
+      i.pt.setMarkdownBodyRef,
+      i.pt.setDataKey
+    ]).pipe(
+      rx.filter(([[, mdKey], [, key]]) => mdKey === key),
+      rx.switchMap(([[, mdKey, dom], [m, key]]) => outputTable.l.topLevelItemIdsUpdated.pipe(
+        actionRelatedToPayload(m.i),
+        rx.take(1),
+        rx.withLatestFrom(outputTable.l.itemById),
+        rx.tap(([, [, itemById]]) => {
+          for (const [id, item] of itemById.entries()) {
+            const textDiv = dom.querySelector('[id="mdt-' + id + '"]');
+            if (textDiv) {
+              item.textDom = textDiv as HTMLDivElement;
+            }
+          }
+        })
+      ))
+    ));
+
+  r('clicked', i.pt.setLayoutControl.pipe(
+    rx.switchMap(([, layout]) => layout.inputTable.l.setFrontLayerRef),
+    rx.filter(([, ref]) => ref != null),
+    rx.switchMap(([, scrollable]) => i.pt.clicked.pipe(
+      rx.switchMap(([, id]) => {
+        return outputTable.l.itemById.pipe(
+          rx.take(1),
+          rx.switchMap(([, map]) => {
+            const itemState = map.get(id);
+            const rect = itemState?.textDom!.getBoundingClientRect();
+            // eslint-disable-next-line no-console
+            console.log(rect);
+            i.dp.togglePopup(false);
+            if (rect)
+              return rx.timer(250).pipe(
+                rx.tap(() => scrollable!.scrollTo({
+                  left: 0,
+                  top: rect.y - scrollable!.getBoundingClientRect().y,
+                  behavior: 'smooth'
+                }))
+              );
+            else
+              return rx.EMPTY;
+          })
+        );
+      })
+    ))
   ));
 
-  r('layout deviceSize', i.pt.setLayoutControl.pipe(
+  r('For desktop device: layout.onTopAppBarScrollChange -> changeFixPosition', i.pt.setLayoutControl.pipe(
+    rx.switchMap(([, layout]) => layout.inputTable.l.setDeviceSize.pipe(
+      // rx.filter(([, size]) => size === 'desktop'),
+      rx.switchMap(([, size]) => {
+        if (size === 'desktop') {
+          return rx.merge(
+            layout.outputTable.l.onTopAppBarScrollChange.pipe(
+              rx.tap(([, outOfViewPort]) => {
+                o.dp.changeFixPosition(outOfViewPort);
+              }),
+              composite.labelError('-> changeFixPosition')
+            )
+          );
+        } else {
+          return rx.EMPTY;
+        }
+      })
+    ))
+  ));
+
+  r('For non-desktop device: -> togglePopup', i.pt.setLayoutControl.pipe(
     rx.switchMap(([, layout]) => layout.inputTable.l.setDeviceSize),
-    rx.map(([, deviceSize]) => {
-    })
+    rx.filter(([, size]) => size !== 'desktop'),
+    rx.switchMap(() => {
+      o.dp.changeFixPosition(false);
+      return i.pt.togglePopup;
+    }),
+    rx.distinctUntilChanged(([, a], [, b]) => a === b),
+    rx.map(([m, on]) => {
+      o.dpf.togglePopup(m, on);
+      return on;
+    }),
+    // eslint-disable-next-line multiline-ternary
+    rx.concatMap(on => on ? rx.concat(
+      rx.defer(() => {o.dp.togglePopupClassName('toggleOnBegin'); return rx.EMPTY; }),
+      rx.timer(16),
+      rx.defer(() => {o.dp.togglePopupClassName('toggleOn'); return rx.EMPTY; }),
+      rx.timer(300)
+    ) : rx.concat(
+      rx.defer(() => {o.dp.togglePopupClassName('toggleOnBegin'); return rx.EMPTY; }),
+      rx.timer(300),
+      rx.defer(() => {o.dp.togglePopupClassName(''); return rx.EMPTY; }),
+      rx.timer(16)
+    ))
   ));
 
-  r('When position changed', o.pt.changeFixPosition.pipe(
+  r('When changeFixPosition', o.pt.changeFixPosition.pipe(
     op.withLatestFrom(
       i.pt.onPlaceHolderRef.pipe(
         rx.map(([, ref]) => ref),
@@ -116,21 +200,27 @@ export function createControl(uiDirtyCheck: (immutableObj: any) => any) {
         op.filter((ref): ref is NonNullable<typeof ref> => ref != null)
       )
     ),
-    op.map(([[, fixed], placeHolderRef, contentRef]) => {
+    op.switchMap(([[, fixed], placeHolderRef, contentRef]) => {
       if (fixed) {
         const w = placeHolderRef.clientWidth + 'px';
         const h = placeHolderRef.clientHeight + 'px';
         placeHolderRef.style.width = w;
         placeHolderRef.style.height = h;
         contentRef.style.width = w;
+        return rx.EMPTY;
       } else {
-        placeHolderRef.style.width = '';
         contentRef.style.width = '';
+        // In Safari, the flash is pretty abvious when "fixed" position change which causes browsr reflow
+        return rx.timer(2000).pipe(
+          rx.tap(() => {
+            placeHolderRef.style.width = '';
+          })
+        );
       }
     })
   ));
 
-  let state: ActionTableDataType<TocUIEvents, typeof tocOutputTableFor> | undefined;
+  let state: TocUIEventTable | undefined;
   r('update state', composite.outputTable.dataChange$.pipe(
     rx.tap(obj => {
       state = obj;